requirements.txt
streamlit>=1.28.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
